Only allow users to delete the latest court session

+++ apps/judicial-system/backend/src/app/modules/repository/services/courtSessionRepository.service.ts
@@ -149,6 +149,18 @@ export class CourtSessionRepositoryService {
 
       const transaction = options?.transaction
 
+      // Only allow users to delete the latest court session
+      const courtSessionsInCase = await this.courtSessionModel.findAll({
+        where: { caseId },
+        order: [['created', 'DESC']],
+      })
+
+      if (courtSessionsInCase[0].id !== courtSessionId) {
+        throw new InternalServerErrorException(
+          `Could not delete court session ${courtSessionId} of case ${caseId}. Court session to delete is not the latest court session in the case.`,
+        )
+      }
+
       // First delete all documents in the session
       await this.courtDocumentRepositoryService.deleteDocumentsInSession(
         caseId,
